backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
itsdangerous==2.2.0
requests==2.32.5
pandas==2.2.2
openpyxl==3.1.5
Pillow==10.4.0
PyPDF2==3.0.1
orjson==3.10.7
beanie==1.26.0
motor==3.3.2
pymongo==4.6.3
pydantic-settings==2.3.4
python-dotenv==1.0.1

# Auth and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.2
google-auth==2.35.0

# ML and processing
scikit-learn==1.5.2
scipy==1.13.1

# Reports/PDF
reportlab==4.2.2

# Low-level bcrypt backend
bcrypt==4.1.2

# RequestsDependencyWarning can trigger if chardet is too new (reportlab pulls chardet)
chardet<6
